+++ code/manage/contact-delete.go
@@ -19,7 +19,9 @@ func ContactRemove(cmd *cobra.Command, args []string) {
 
 		id, err := utils.UfS(x)
 		if err != nil {
-			logrus.Fatal(err)
+			// logrus.Fatal(err)
+			logrus.Warn(err)
+			break
 		}
 
 		contact := db.Contact{}